% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cptac-luad-data.R
\docType{data}
\name{CPTAC_LUAD}
\alias{CPTAC_LUAD}
\title{Protein Abundance data from LUAD}
\format{
A list with 'tissue' and 'tumor' elements, each a data.frame object.
}
\source{
Downloaded from CPTAC
}
\usage{
data(CPTAC_LUAD)
}
\description{
Protein abundance data downloaded from Clinical Proteomic Tumor Analysis Consortium (CPTAC)
for lung adenocarcinoma (LUAD). Data contains protein abundances for 10,316 proteins in
101 normal tissue samples and 110 tumor samples. Proteins are in rows and samples are in columns.
Protein are labeled with Hugo IDs.
}
\examples{
data(CPTAC_LUAD)
\donttest(dim(CPTAC_LUAD))
}
\keyword{datasets}
