% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichmentAnalysisPlot.R
\name{EnrichmentAnalysisPlot}
\alias{EnrichmentAnalysisPlot}
\title{Enrichment Analysis Plotting function}
\usage{
EnrichmentAnalysisPlot(
  cur.ges,
  cur.regul,
  cur.minsize = 30,
  cur.score = 0,
  cur.seed = 1,
  non.parametric = TRUE,
  cur.title = "Non-parametric Analytical Rank-based Enrichment Analysis",
  cur.subtitle = NULL,
  cur.x.lab = "Gene Expression Signature",
  cur.y.lab = "KS Running Sum Statistic",
  title.font.size = 22,
  subtitle.font.size = 20,
  axis.title.font.size = 18,
  axis.text.font.size = 15,
  positive.color = "red1",
  middle.color = "grey40",
  negative.color = "blue1",
  enrich.lwd = 1,
  mid.bar.lwd = 1,
  black.bar.lwd = 0.75,
  sig.lwd = 0.85
)
}
\arguments{
\item{cur.ges}{Gene expression signature (named numeric vector)}

\item{cur.regul}{Gene set parameterized with Regulation Confidence (likelihood) and Mode of Regulation (tfmode)}

\item{cur.minsize}{Minimum number of gene set members that are present in the gene expression signature. Default of 30.}

\item{cur.seed}{Random number generator seed to ensure ties in the gene expression signature are broken in a reproducible manner. Default of 1.}

\item{cur.title}{The text for the main title.}

\item{cur.subtitle}{The text for the subtitle. Default is NULL. If non-NULL, this will replace the output NaRnEA PES, NES, and two-sided p-value for the enrichment of the gene set in the gene expression signature.}

\item{cur.x.lab}{The text for the x-axis label.}

\item{cur.y.lab}{The text for the y-axis label.}

\item{title.font.size}{The font size for the main title.}

\item{subtitle.font.size}{The font size for the subtitle.}

\item{axis.title.font.size}{The font size for the x-axis label and the y-axis label.}

\item{axis.text.font.size}{The font size for the text of the x-axis and y-axis.}

\item{positive.color}{The color for the plot elements corresponding to gene set members with a positive Mode of Regulation (tfmode).}

\item{middle.color}{The middle color for the color gradient between the positive color and the negative color.}

\item{negative.color}{The color for the plot elements corresponding to gene set members with a negative Mode of Regulation (tfmode)}

\item{enrich.lwd}{The linewidth for the running sum statistic visualization.}

\item{mid.bar.lwd}{The linewidth for the horizontal bar denoting the expected value of the running sum statistic.}

\item{black.bar.lwd}{The linewidth for the vertical bar denoting where the gene expression signature switches from negative values to positive values.}

\item{sig.lwd}{The linewidth for the vertical lines denoting the location of each gene set member in the gene expression signature.}
}
\value{
The NaRnEA PES, NES, and two-sided p-value for the enrichment of the gene set in the gene expression signature as well as a figure made with ggplot2 that visualizes the enrichment of the gene set in the gene expression signature using the two-sample Kolmogorov Smirnov running sum statistic (if cur.score = 0) or the GSEA running sum statistic (if cur.score = 1).
}
\description{
Enrichment Analysis Plotting function
}
