% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichmentPlot.R
\name{EnrichmentPlot}
\alias{EnrichmentPlot}
\title{Nonparametric analytical Rank-based Enrichment Analysis plotting function}
\usage{
EnrichmentPlot(
  signature,
  association.weight,
  association.mode,
  minimum.size = 30,
  seed = 1,
  title.text = "Nonparametric analytical Rank-based Enrichment Analysis",
  x.label.text = "Nonparametric Gene Expression Signature",
  line.width = 1,
  lower.color = "blue1",
  middle.color = "grey50",
  upper.color = "red1",
  conf.int = TRUE,
  conf.level = 0.95,
  boot.num = 100,
  sig.figs = 4
)
}
\arguments{
\item{signature}{Gene expression signature (named numeric vector)}

\item{association.weight}{Association Weight values for gene set members (named numeric vector)}

\item{association.mode}{Association Mode values for gene set members (named numeric vector)}

\item{minimum.size}{Minimum number of gene set members. Default of 30.}

\item{seed}{Random number generator seed to ensure reproducibility manner. Default of 1.}

\item{title.text}{Text for title of the plot.}

\item{x.label.text}{Text for x-axis label of the plot.}

\item{line.width}{Line width for the gene set member segments in the nonparametric gene expression signature.}

\item{lower.color}{Color for the gene set members with negative Association Mode values.}

\item{middle.color}{Color for the gene set members with Association Mode values close to zero.}

\item{upper.color}{Color for the gene set members with positive Association Mode values.}

\item{conf.int}{Flag to compute the confidence interval for the NaRnEA Proportional Enrichment Score using a Fisher Transformation and gene set member sampling with replacement.}

\item{conf.level}{Desired coverage of the confidence interval for the NaRnEA Proportional Enrichment Score. Default of 0.95.}

\item{boot.num}{Number of times to sample the gene set members with replacement when estimating the confidence interval for the NaRnEA Proportional Enrichment Score. Default of 100.}

\item{sig.figs}{Number of significant figures to display when plotting numerical values. Default of 4.}
}
\value{
This function returns the NaRnEA Proportional Enrichment Score (PES) with a confidence interval (if desired), the NaRnEA Normalized Enrichment Score (NES), the NaRnEA two-sided p-value of the NaRnEA, and a figure made with ggplot2 that visualizes the enrichment of the gene set members in the nonparametrically transformed differential gene expression signature.
}
\description{
Nonparametric analytical Rank-based Enrichment Analysis plotting function
}
