% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene-name-map-data.R
\docType{data}
\name{gene.name.map}
\alias{gene.name.map}
\title{Gene name convention map}
\format{
A data.frame
}
\source{
Generated using biomaRt
}
\usage{
data(gene.name.map)
}
\description{
A data frame containing gene names across five conventions: Ensembl transcripts (enst.values),
Ensembl genes (ensg.values), Entrez ID (entrez.values), Hugo gene names (hugo,values), and Biotype
values (biotype.values). Contains 177,680 rows.
}
\examples{
data(gene.name.map)
\donttest{dim(gene.name.map)}
}
\keyword{datasets}
