% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luad-network-data.R
\docType{data}
\name{LUAD_network}
\alias{LUAD_network}
\title{LUAD Network List}
\format{
Two objects: \code{LUAD.regulon.list}, a list of regulon objects, each with parametrized network targets generated from ARACNe3;
and \code{LUAD.mwu.ges}, a vector of GES values for the tumor-vs-normal GES.
}
\source{
Objects generated from TCGA LUAD data.
}
\usage{
data(LUAD_network)
}
\description{
Network list generated from tumor samples using ARACNe3.
This object can be used to run the simplified vignette in the NaRnEA README.
For full instructions on how to generate this object, see the full LUAD vignette.
Genes are labeled with Entrez IDs.
}
\examples{
data(LUAD_network)
\donttest{dim(LUAD_network)}
}
\keyword{datasets}
