% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_NaRnEA.R
\name{matrixNaRnEA}
\alias{matrixNaRnEA}
\title{Runs the NaRnEA algorithm on multiple samples / regulons simultaneously.}
\usage{
matrixNaRnEA(ges.mat, regulon.list, seed.val = 343, min.targets = 30)
}
\arguments{
\item{ges.mat}{Matrix of gene expression signatures (genes X samples).}

\item{regulon.list}{List of regulon lists, with am and aw values.}

\item{seed.val}{Value of seed to ensure reproducibility. Default of 343.}

\item{min.targets}{Minimum number of targets needed between a regulon and a GES. Default of 30.}
}
\value{
List of matrices; 'nes' and 'pes'.
}
\description{
Runs the NaRnEA algorithm on multiple samples / regulons simultaneously.
}
