% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga-coad-data.R
\docType{data}
\name{TCGA_COAD}
\alias{TCGA_COAD}
\title{Bulk RNA-Seq data from COAD}
\format{
A matrix.
}
\source{
Downloaded using TCGAbiolinks
}
\usage{
data(TCGA_COAD)
}
\description{
Bulk RNA-Seq data from The Cancer Genome Atlas (TCGA) for colon adenocarcinoma (COAD)
with HT-Seq counts data for 19,350 genes in 519 samples. Genes are in rows and samples
are in columns. Genes are labeled with Entrez IDs.
}
\examples{
data(TCGA_COAD)
\donttest(dim(TCGA_COAD))
}
\keyword{datasets}
