% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga-hnsc-data.R
\docType{data}
\name{TCGA_HNSC}
\alias{TCGA_HNSC}
\title{Bulk RNA-Seq data from HNSC}
\format{
A matrix.
}
\source{
Downloaded using TCGAbiolinks
}
\usage{
data(TCGA_HNSC)
}
\description{
Bulk RNA-Seq data from The Cancer Genome Atlas (TCGA) for head and neck squamous cell carcinoma (HNSC)
with HT-Seq counts data for 19,350 genes in 543 samples. Genes are in rows and samples
are in columns. Genes are labeled with Entrez IDs.
}
\examples{
data(TCGA_HNSC)
\donttest{dim(TCGA_HNSC)}
}
\keyword{datasets}
