% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga-luad-data.R
\docType{data}
\name{TCGA_LUAD}
\alias{TCGA_LUAD}
\title{Bulk RNA-Seq data from LUAD}
\format{
A matrix.
}
\source{
Downloaded using TCGAbiolinks
}
\usage{
data(TCGA_LUAD)
}
\description{
Bulk RNA-Seq data from The Cancer Genome Atlas (TCGA) for lung adenocarcinoma (LUAD)
with HT-Seq counts data for 19,350 genes in 590 samples Genes are in rows and samples
are in columns. Genes are labeled with Entrez IDs.
}
\examples{
data(TCGA_LUAD)
\donttest(dim(TCGA_LUAD))
}
\keyword{datasets}
