% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.R
\name{GSEA}
\alias{GSEA}
\title{GSEA as described in Subramanian et. al. 2005 PNAS}
\usage{
GSEA(ges, geneset, score = 1)
}
\arguments{
\item{ges}{Gene expression signature (named numeric vector)}

\item{geneset}{Set of genes to test for enrichment in the GES (character vector)}

\item{score}{Power score for GSEA. Default value of 1 (standard for GSEA; use 0 for two-sample KS test)}
}
\value{
An enrichment score for geneset in the GES.
}
\description{
GSEA as described in Subramanian et. al. 2005 PNAS
}
