% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NaRnEA.R
\name{NaRnEA}
\alias{NaRnEA}
\title{NaRnEA (Non-parameetric analytical rank-based enrichment analysis) function}
\usage{
NaRnEA(ges, regulon, minsize = 30, seed = 1, leading.edge = FALSE)
}
\arguments{
\item{ges}{Gene expression signature (named numeric vector)}

\item{regulon}{Gene set parameterized w/ Regulation Confidence (likelihood) and Mode of Regulation (tfmode)}

\item{minsize}{Minimum overlap between regulon and gene expression signature. Default of 30.}

\item{seed}{Randomly generated seed to ensure ties in the GES are broken in a reproducible manner. Default of 1.}

\item{leading.edge}{Flag to compute leading edge z-scores for the members of the gene set. False by default.}
}
\value{
A list with the NES, PES, maximum and minimum possible NES, and the leading edge if specified.
}
\description{
NaRnEA (Non-parameetric analytical rank-based enrichment analysis) function
}
