% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NaRnEA.R
\name{NaRnEA}
\alias{NaRnEA}
\title{Nonparametric analytical rank-based enrichment analysis (NaRnEA) function}
\usage{
NaRnEA(
  signature,
  association.weight,
  association.mode,
  ledge = TRUE,
  minimum.size = 30,
  seed = 1
)
}
\arguments{
\item{signature}{Gene expression signature (named numeric vector)}

\item{association.weight}{Association Weight values for gene set members (named numeric vector)}

\item{association.mode}{Association Mode values for gene set members (named numeric vector)}

\item{ledge}{Flag to compute ledge p-values for gene set members.}

\item{minimum.size}{Minimum number of gene set members. Default of 30.}

\item{seed}{Random number generator seed to ensure reproducibility manner. Default of 1.}
}
\value{
A list with the Proportional Enrichment Score (PES), Normalized Enrichment Score (NES), and the leading edge p-values (NA if ledge set to FALSE).
}
\description{
Nonparametric analytical rank-based enrichment analysis (NaRnEA) function
}
